requirements.txt
click==8.1.3
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
